Bump unzipper and nodejs versions (#250)
[Lokalise pulls have been failing for a few days now with a "Premature
close" error](https://github.com/color/color/actions/runs/9845697846).
After a bit of digging, I found [this issue in the unzipper package
repo](https://github.com/ZJONSSON/node-unzipper/issues/269) describing
the same error that we’re getting. We’re [supposedly running this
action](https://github.com/color/lokalise-actions/blob/main/.tool-versions)
on node 16 so I didn’t think that issue would apply. However, I saw a
warning on the CI failure linking to [this
changelog](https://github.blog/changelog/2024-03-07-github-actions-all-actions-will-run-on-node20-instead-of-node16-by-default/)
stating that the action would start running on node 20 by default.

In this PR I'm bumping us to the most recent version of both nodejs and
unzipper. This fixes tests locally, and hopefully allows us to pull from
lokalise again (without breaking pushes 😁).

+++ src/lokalise/download-files.ts
@@ -1,7 +1,7 @@
 import * as core from '@actions/core';
 
 import { DownloadFileParams, LokaliseApi } from '@lokalise/node-api';
-import { createReadStream, createWriteStream } from 'fs';
+import { createReadStream, createWriteStream, existsSync } from 'fs';
 import { unlink, writeFile } from 'fs/promises';
 import got from 'got';
 import { join } from 'path';
@@ -137,7 +137,9 @@ async function extractBundleToDownloadDirectory(
     }
   }
 
-  await unlink(bundleDestination);
+  if (existsSync(bundleDestination)) {
+    await unlink(bundleDestination);
+  }
 }
 
 /**
